refactor(test): BaseServiceTest 분리
